set up launcher commands
